Simple testbenches on write resp bank
