+++ rtl/simmem_pkg.sv
@@ -7,16 +7,6 @@
 package simmem_pkg;
 
   typedef enum logic {
-    READ_REQ = 1'b0,
-    WRITE_RESP = 1'b1
-  } data_channel_e;
-
-  typedef enum logic {
-    READ_DATA = 1'b0,
-    WRITE_RESP = 1'b1
-  } bank_channel_e;
-
-  typedef enum logic {
     STRUCT_RAM = 1'b0,
     NEXT_ELEM_RAM = 1'b1
   } ram_bank_e;
