+++ simmem.core
@@ -10,22 +10,30 @@ filesets:
     depend:
       - lowrisc:prim_generic:ram_2p
     files:
-      - rtl/simmem_message_bank.sv
+      - rtl/simmem_pkg.sv
+      - rtl/simmem_write_resp_bank.sv
     file_type: systemVerilogSource
 
-  files_dv_message_bank:
+  files_rtl_write_resp_bank:
+    depend:
+      - lowrisc:prim_generic:ram_2p
     files:
-      - dv/simmem_message_bank/cpp/simmem_message_bank_tb.h
-      - dv/simmem_message_bank/cpp/simmem_message_bank_tb.cc
+      - rtl/simmem_pkg.sv
+      - rtl/simmem_write_resp_bank.sv
+    file_type: systemVerilogSource
+
+  files_dv_write_resp_bank:
+    files:
+      - dv/simmem_write_resp_bank/cpp/simmem_write_resp_bank_tb.cc
     file_type: cppSource
 
 targets:
-  sim_message_bank:
+  sim_write_resp_bank:
     default_tool: verilator
     filesets:
-      - files_rtl
-      - files_dv_message_bank
-    toplevel: simmem_message_bank_tb
+      - files_rtl_write_resp_bank
+      - files_dv_write_resp_bank
+    toplevel: simmem_write_resp_bank
     tools:
       verilator:
         mode: cc
@@ -42,7 +50,7 @@ targets:
 # -O
 #   Optimization levels have a large impact on the runtime performance of the
 #   simulation model. -O2 and -O3 are pretty similar, -Os is slower than -O2/-O3
-          - '-CFLAGS "-std=c++11 -Wall -DVM_TRACE_FMT_FST -DTOPLEVEL_NAME=simmem_message_bank_tb -g -O0"'
+          - '-CFLAGS "-std=c++11 -Wall -DVM_TRACE_FMT_FST -DTOPLEVEL_NAME=simmem_write_resp_bank_tb -g -O0"'
           - '-LDFLAGS "-pthread -lutil"'
           - "-Wall"
           - "-Wno-PINCONNECTEMPTY"
